+++ app/src/main/java/sleepless_nights/location_alarm/geofence/use_cases/GeofenceBroadcastReceiver.java
@@ -39,7 +39,7 @@ public class GeofenceBroadcastReceiver extends BroadcastReceiver {
                     Log.wtf(TAG, "Got geofence with null id");
                     continue;
                 }
-                Integer alarmId = GeofenceRepository.getInstance(context)
+                Long alarmId = GeofenceRepository.getInstance(context)
                         .getAlarmIdByGeofenceId(geofenceId);
                 if (alarmId == null) {
                     continue; //Triggered unregistered geofence
@@ -66,7 +66,7 @@ public class GeofenceBroadcastReceiver extends BroadcastReceiver {
         }
     }
 
-    private void doAlarm(Context context, int alarmId) {
+    private void doAlarm(Context context, long alarmId) {
         Intent intent = new Intent(AlarmService.ACTION_DO_ALARM);
         intent.putExtra(AlarmService.INTENT_EXTRA_ALARM_ID, alarmId);
         context.getApplicationContext().startService(intent);
